Target and source paths are now required on the command line.
The path to the target and source path can no longer be specified in the layout JSON file, but must be specified on the program command line.

Note, the target path does not include the final directory name, as this is actually given as the folder for the first page in the "site": property.